requirements.txt
PySide6>=6.7
python-docx>=1.1.2
Pillow>=10.0
easyocr>=1.7.1
